remove nestjs
