+++ .erb/configs/webpack.config.base.ts
@@ -58,8 +58,7 @@ const configuration: webpack.Configuration = {
 
   plugins: [
     new webpack.IgnorePlugin({
-      resourceRegExp:
-        /^@nestjs\/(microservices|socket|websockets|platform-express)/,
+      resourceRegExp: /^@nestjs\/(microservices|socket|websockets|platform-express)/,
     }),
     new webpack.EnvironmentPlugin({
       NODE_ENV: 'production',
