+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "sanita-ticket-js",
-  "version": "1.6.0",
+  "version": "1.7.0",
   "description": "Electron app to read 2003 mdb files and create a purchase ticket.",
   "keywords": [
     "electron",
@@ -92,8 +92,6 @@
     "@fontsource/inter": "^5.0.8",
     "@mui/icons-material": "^5.14.11",
     "@mui/joy": "^5.0.0-beta.8",
-    "@nestjs/common": "^10.2.6",
-    "@nestjs/core": "^10.2.6",
     "@supabase/supabase-js": "^2.43.4",
     "buffer": "^6.0.3",
     "dotenv": "^16.4.5",
